Pull out target generation from the analyzer, cleanup type annotations
