+++ lib/cocoapods/installer/analyzer.rb
@@ -16,12 +16,9 @@ module Pod
       autoload :PodVariantSet,             'cocoapods/installer/analyzer/pod_variant_set'
       autoload :SandboxAnalyzer,           'cocoapods/installer/analyzer/sandbox_analyzer'
       autoload :SpecsState,                'cocoapods/installer/analyzer/specs_state'
+      autoload :TargetGenerator,           'cocoapods/installer/analyzer/target_generator'
       autoload :TargetInspectionResult,    'cocoapods/installer/analyzer/target_inspection_result'
       autoload :TargetInspector,           'cocoapods/installer/analyzer/target_inspector'
-
-      # @return [String] The version of iOS which requires binaries with only 64-bit architectures
-      #
-      IOS_64_BIT_ONLY_VERSION = Version.new('11.0')
 
       # @return [Sandbox] The sandbox to use for this analysis.
       #
@@ -88,7 +85,7 @@ module Pod
         validate_podfile!
         validate_lockfile_version!
         if installation_options.integrate_targets?
-          target_inspections = inspect_targets_to_integrate
+          target_inspections = inspect_targets_to_integrate(@podfile_dependency_cache.target_definition_list)
         else
           verify_platforms_specified!
           target_inspections = {}
@@ -183,8 +180,8 @@ module Pod
         pods_to_update != nil
       end
 
-      # @return [Symbol] Whether and how the dependencies in the Podfile
-      #                  should be updated.
+      # @return [:none, :all, :selected] Whether and how the dependencies in the Podfile
+      #         should be updated.
       #
       def update_mode
         if !pods_to_update
@@ -196,6 +193,8 @@ module Pod
         end
       end
 
+      # @return [Array<Pod::Dependency>] All the dependencies in the podfile
+      #
       def podfile_dependencies
         @podfile_dependency_cache.podfile_dependencies
       end
@@ -230,14 +229,12 @@ module Pod
       # Compares the {Podfile} with the {Lockfile} in order to detect which
       # dependencies should be locked.
       #
-      # @return [SpecsState] the states of the Podfile specs.
-      #
       # @note   As the target definitions share the same sandbox they should have
       #         the same version of a Pod. For this reason this method returns
       #         the name of the Pod (root name of the dependencies) and doesn't
       #         group them by target definition.
       #
-      # @return [SpecState]
+      # @return [SpecsState] the states of the Podfile specs.
       #
       def generate_podfile_state
         if lockfile
@@ -388,26 +385,17 @@ module Pod
       # Creates the models that represent the targets generated by CocoaPods.
       #
       # @param  [Hash{Podfile::TargetDefinition => Array<ResolvedSpecification>}] resolver_specs_by_target
-      #         mapping of targets to resolved specs (containing information about test usage)
-      #         aggregate targets
+      #         mapping of target definition to resolved specs (containing information about test usage)
       #
-      # @param  [Array<TargetInspection>] target_inspections
+      # @param  [Hash{Podfile::TargetDefinition => TargetInspectionResult>] target_inspections
       #         the user target inspections used to construct the aggregate and pod targets.
       #
       # @return [Array<AggregateTarget>] the list of aggregate targets generated.
       #
       def generate_targets(resolver_specs_by_target, target_inspections)
-        resolver_specs_by_target = resolver_specs_by_target.reject { |td, _| td.abstract? }
-        pod_targets = generate_pod_targets(resolver_specs_by_target, target_inspections)
-        aggregate_targets = resolver_specs_by_target.keys.map do |target_definition|
-          generate_target(target_definition, target_inspections, pod_targets, resolver_specs_by_target)
-        end
-        aggregate_targets.each do |target|
-          search_paths_aggregate_targets = aggregate_targets.select do |aggregate_target|
-            target.target_definition.targets_to_inherit_search_paths.include?(aggregate_target.target_definition)
-          end
-          target.search_paths_aggregate_targets.concat(search_paths_aggregate_targets).freeze
-        end
+        aggregate_targets = TargetGenerator.generate_aggregate_targets(resolver_specs_by_target, target_inspections,
+                                                                       @podfile_dependency_cache, sandbox,
+                                                                       installation_options)
         if installation_options.integrate_targets?
           # Copy embedded target pods that cannot have their pods embedded as frameworks to
           # their host targets, and ensure we properly link library pods to their host targets
@@ -431,49 +419,8 @@ module Pod
         aggregate_targets
       end
 
-      # Setup the aggregate target for a single user target
+      # @param  [Array<AggregateTarget>] aggregate_targets
       #
-      # @param  [TargetDefinition] target_definition
-      #         the target definition for the user target.
-      #
-      # @param  [Array<TargetInspection>] target_inspections
-      #         the user target inspections used to construct the aggregate and pod targets.
-      #
-      # @param  [Array<PodTarget>] pod_targets
-      #         the pod targets, which were generated.
-      #
-      # @param  [Hash{Podfile::TargetDefinition => Array<ResolvedSpecification>}] resolver_specs_by_target
-      #         the resolved specifications grouped by target.
-      #
-      # @return [AggregateTarget]
-      #
-      def generate_target(target_definition, target_inspections, pod_targets, resolver_specs_by_target)
-        target_requires_64_bit = requires_64_bit_archs?(target_definition.platform)
-        if installation_options.integrate_targets?
-          target_inspection = target_inspections[target_definition]
-          raise "missing inspection: #{target_definition.name}" unless target_inspection
-          user_project = target_inspection.project
-          client_root = target_inspection.client_root
-          user_target_uuids = target_inspection.project_target_uuids
-          user_build_configurations = target_inspection.build_configurations
-          archs = target_requires_64_bit ? ['$(ARCHS_STANDARD_64_BIT)'] : target_inspection.archs
-        else
-          user_project = nil
-          client_root = config.installation_root.realpath
-          user_target_uuids = []
-          user_build_configurations = target_definition.build_configurations || Target::DEFAULT_BUILD_CONFIGURATIONS
-          archs = target_requires_64_bit ? ['$(ARCHS_STANDARD_64_BIT)'] : []
-        end
-        platform = target_definition.platform
-        build_configurations = user_build_configurations.keys.concat(target_definition.all_whitelisted_configurations).uniq
-        pod_targets_for_build_configuration = filter_pod_targets_for_target_definition(target_definition, pod_targets,
-                                                                                       resolver_specs_by_target,
-                                                                                       build_configurations)
-        AggregateTarget.new(sandbox, target_definition.uses_frameworks?, user_build_configurations, archs, platform,
-                            target_definition, client_root, user_project, user_target_uuids,
-                            pod_targets_for_build_configuration)
-      end
-
       # @return [Array<PodTarget>] The model representations of pod targets.
       #
       def calculate_pod_targets(aggregate_targets)
@@ -491,62 +438,6 @@ module Pod
         (aggregate_target_pod_targets + test_dependent_targets + app_dependent_targets).uniq
       end
 
-      # Returns a filtered list of pod targets that should or should not be part of the target definition. Pod targets
-      # used by tests only are filtered.
-      #
-      # @param [TargetDefinition] target_definition
-      #        the target definition to use as the base for filtering
-      #
-      # @param [Array<PodTarget>] pod_targets
-      #        the array of pod targets to check against
-      #
-      # @param  [Hash{Podfile::TargetDefinition => Array<ResolvedSpecification>}] resolver_specs_by_target
-      #         the resolved specifications grouped by target.
-      #
-      # @param  [Array<String>] build_configurations
-      #         The list of all build configurations the targets will be built for.
-      #
-      # @return [Hash<String => Array<PodTarget>>]
-      #         the filtered list of pod targets, grouped by build configuration.
-      #
-      def filter_pod_targets_for_target_definition(target_definition, pod_targets, resolver_specs_by_target, build_configurations)
-        pod_targets_by_build_config = Hash.new([].freeze)
-        build_configurations.each { |config| pod_targets_by_build_config[config] = [] }
-
-        pod_targets.each do |pod_target|
-          next unless pod_target.target_definitions.include?(target_definition)
-          next unless resolver_specs_by_target[target_definition].any? do |resolver_spec|
-            !resolver_spec.used_by_non_library_targets_only? && pod_target.specs.include?(resolver_spec.spec)
-          end
-
-          pod_name = pod_target.pod_name
-
-          dependencies = @podfile_dependency_cache.target_definition_dependencies(target_definition).select do |dependency|
-            Specification.root_name(dependency.name) == pod_name
-          end
-
-          build_configurations.each do |configuration_name|
-            whitelists = dependencies.map do |dependency|
-              target_definition.pod_whitelisted_for_configuration?(dependency.name, configuration_name)
-            end.uniq
-
-            case whitelists
-            when [], [true] then nil
-            when [false] then next
-            else
-              raise Informative, "The subspecs of `#{pod_name}` are linked to " \
-                "different build configurations for the `#{target_definition}` " \
-                'target. CocoaPods does not currently support subspecs across ' \
-                'different build configurations.'
-            end
-
-            pod_targets_by_build_config[configuration_name] << pod_target
-          end
-        end
-
-        pod_targets_by_build_config
-      end
-
       # Setup the pod targets for an aggregate target. Deduplicates resulting
       # targets by grouping by platform and subspec by their root
       # to create a {PodTarget} for each spec.
@@ -560,203 +451,7 @@ module Pod
       # @return [Array<PodTarget>]
       #
       def generate_pod_targets(resolver_specs_by_target, target_inspections)
-        if installation_options.deduplicate_targets?
-          distinct_targets = resolver_specs_by_target.each_with_object({}) do |dependency, hash|
-            target_definition, dependent_specs = *dependency
-            dependent_specs.group_by(&:root).each do |root_spec, resolver_specs|
-              all_specs = resolver_specs.map(&:spec)
-              all_specs_by_type = all_specs.group_by(&:spec_type)
-              library_specs = all_specs_by_type[:library] || []
-              test_specs = all_specs_by_type[:test] || []
-              app_specs = all_specs_by_type[:app] || []
-              pod_variant = PodVariant.new(library_specs, test_specs, app_specs, target_definition.platform, target_definition.uses_frameworks?)
-              hash[root_spec] ||= {}
-              (hash[root_spec][pod_variant] ||= []) << target_definition
-              pod_variant_spec = hash[root_spec].keys.find { |k| k == pod_variant }
-              pod_variant_spec.test_specs.concat(test_specs).uniq!
-              pod_variant_spec.app_specs.concat(app_specs).uniq!
-            end
-          end
-
-          pod_targets = distinct_targets.flat_map do |_root, target_definitions_by_variant|
-            suffixes = PodVariantSet.new(target_definitions_by_variant.keys).scope_suffixes
-            target_definitions_by_variant.flat_map do |variant, target_definitions|
-              generate_pod_target(target_definitions, target_inspections, variant.specs + variant.test_specs + variant.app_specs, :scope_suffix => suffixes[variant])
-            end
-          end
-
-          all_resolver_specs = resolver_specs_by_target.values.flatten.map(&:spec).uniq
-          pod_targets_by_name = pod_targets.group_by(&:pod_name).each_with_object({}) do |(name, values), hash|
-            # Sort the target by the number of activated subspecs, so that
-            # we prefer a minimal target as transitive dependency.
-            hash[name] = values.sort_by { |pt| pt.specs.count }
-          end
-          pod_targets.each do |target|
-            all_specs = all_resolver_specs.group_by(&:name)
-            dependencies = dependencies_for_specs(target.library_specs.to_set, target.platform, all_specs.dup).group_by(&:root)
-            target.dependent_targets = filter_dependencies(dependencies, pod_targets_by_name, target)
-            target.test_dependent_targets_by_spec_name = target.test_specs.each_with_object({}) do |test_spec, hash|
-              test_dependencies = dependencies_for_specs([test_spec], target.platform, all_specs).group_by(&:root)
-              test_dependencies.delete_if { |k| dependencies.key? k }
-              hash[test_spec.name] = filter_dependencies(test_dependencies, pod_targets_by_name, target)
-            end
-
-            target.app_dependent_targets_by_spec_name = target.app_specs.each_with_object({}) do |app_spec, hash|
-              app_dependencies = dependencies_for_specs([app_spec], target.platform, all_specs).group_by(&:root)
-              app_dependencies.delete_if { |k| dependencies.key? k }
-              hash[app_spec.name] = filter_dependencies(app_dependencies, pod_targets_by_name, target)
-            end
-          end
-        else
-          dedupe_cache = {}
-          resolver_specs_by_target.flat_map do |target_definition, specs|
-            grouped_specs = specs.group_by(&:root).values.uniq
-            pod_targets = grouped_specs.flat_map do |pod_specs|
-              generate_pod_target([target_definition], target_inspections, pod_specs.map(&:spec)).scoped(dedupe_cache)
-            end
-
-            pod_targets.each do |target|
-              all_specs = specs.map(&:spec).group_by(&:name)
-              dependencies = dependencies_for_specs(target.library_specs.to_set, target.platform, all_specs.dup).group_by(&:root)
-              target.dependent_targets = pod_targets.reject { |t| dependencies[t.root_spec].nil? }
-              target.test_dependent_targets_by_spec_name = target.test_specs.each_with_object({}) do |test_spec, hash|
-                test_dependencies = dependencies_for_specs(target.test_specs.to_set, target.platform, all_specs.dup).group_by(&:root)
-                test_dependencies.delete_if { |k| dependencies.key? k }
-                hash[test_spec.name] = pod_targets.reject { |t| test_dependencies[t.root_spec].nil? }
-              end
-              target.app_dependent_targets_by_spec_name = target.app_specs.each_with_object({}) do |app_spec, hash|
-                app_dependencies = dependencies_for_specs(target.app_specs.to_set, target.platform, all_specs.dup).group_by(&:root)
-                app_dependencies.delete_if { |k| dependencies.key? k }
-                hash[app_spec.name] = pod_targets.reject { |t| app_dependencies[t.root_spec].nil? }
-              end
-            end
-          end
-        end
-      end
-
-      def filter_dependencies(dependencies, pod_targets_by_name, target)
-        dependencies.map do |root_spec, deps|
-          pod_targets_by_name[root_spec.name].find do |t|
-            next false if t.platform.symbolic_name != target.platform.symbolic_name ||
-                t.requires_frameworks? != target.requires_frameworks?
-            spec_names = t.specs.map(&:name)
-            deps.all? { |dep| spec_names.include?(dep.name) }
-          end
-        end
-      end
-
-      # Returns the specs upon which the given specs _directly_ depend.
-      #
-      # @note: This is implemented in the analyzer, because we don't have to
-      #        care about the requirements after dependency resolution.
-      #
-      # @param  [Array<Specification>] specs
-      #         The specs, whose dependencies should be returned.
-      #
-      # @param  [Platform] platform
-      #         The platform for which the dependencies should be returned.
-      #
-      # @param  [Hash<String, Specification>] all_specs
-      #         All specifications which are installed alongside.
-      #
-      # @return [Array<Specification>]
-      #
-      def dependencies_for_specs(specs, platform, all_specs)
-        return [] if specs.empty? || all_specs.empty?
-
-        dependent_specs = Set.new
-
-        specs.each do |s|
-          s.dependencies(platform).each do |dep|
-            all_specs[dep.name].each do |spec|
-              dependent_specs << spec
-            end
-          end
-        end
-
-        dependent_specs - specs
-      end
-
-      # Create a target for each spec group
-      #
-      # @param  [Array<TargetDefinition>] target_definitions
-      #         the target definitions of the aggregate target
-      #
-      # @param  [Array<TargetInspection>] target_inspections
-      #         the user target inspections used to construct the aggregate and pod targets.
-      #
-      # @param  [Array<Specification>] specs
-      #         the specifications of an equal root.
-      #
-      # @param  [String] scope_suffix
-      #         @see PodTarget#scope_suffix
-      #
-      # @return [PodTarget]
-      #
-      def generate_pod_target(target_definitions, target_inspections, specs, scope_suffix: nil)
-        target_requires_64_bit = target_definitions.all? { |td| requires_64_bit_archs?(td.platform) }
-        if installation_options.integrate_targets?
-          target_inspections = target_inspections.select { |t, _| target_definitions.include?(t) }.values
-          user_build_configurations = target_inspections.map(&:build_configurations).reduce({}, &:merge)
-          archs = if target_requires_64_bit
-                    ['$(ARCHS_STANDARD_64_BIT)']
-                  else
-                    target_inspections.flat_map(&:archs).compact.uniq.sort
-                  end
-        else
-          user_build_configurations = {}
-          archs = target_requires_64_bit ? ['$(ARCHS_STANDARD_64_BIT)'] : []
-        end
-        host_requires_frameworks = target_definitions.any?(&:uses_frameworks?)
-        platform = determine_platform(specs, target_definitions, host_requires_frameworks)
-        file_accessors = create_file_accessors(specs, platform)
-        PodTarget.new(sandbox, host_requires_frameworks, user_build_configurations, archs, platform, specs,
-                      target_definitions, file_accessors, scope_suffix)
-      end
-
-      # Creates the file accessors for a given pod.
-      #
-      # @param [Array<Specification>] specs
-      #        the specs to map each file accessor to.
-      #
-      # @param [Platform] platform
-      #        the platform to use when generating each file accessor.
-      #
-      # @return [Array<FileAccessor>]
-      #
-      def create_file_accessors(specs, platform)
-        name = specs.first.name
-        pod_root = sandbox.pod_dir(name)
-        path_list = Sandbox::PathList.new(pod_root)
-        specs.map do |spec|
-          Sandbox::FileAccessor.new(path_list, spec.consumer(platform))
-        end
-      end
-
-      # Calculates and returns the platform to use for the given list of specs and target definitions.
-      #
-      # @param [Array<Specification>] specs
-      #        the specs to inspect and calculate the platform for.
-      #
-      # @param [Array<TargetDefinition>] target_definitions
-      #        the target definitions these specs are part of.
-      #
-      # @param [Boolean] host_requires_frameworks
-      #        whether the platform is calculated for a target that needs to be packaged as a framework.
-      #
-      # @return [Platform]
-      #
-      def determine_platform(specs, target_definitions, host_requires_frameworks)
-        platform_name = target_definitions.first.platform.name
-        default = Podfile::TargetDefinition::PLATFORM_DEFAULTS[platform_name]
-        deployment_target = specs.map do |spec|
-          Version.new(spec.deployment_target(platform_name) || default)
-        end.max
-        if platform_name == :ios && host_requires_frameworks
-          minimum = Version.new('8.0')
-          deployment_target = [deployment_target, minimum].max
-        end
-        Platform.new(platform_name, deployment_target)
+        TargetGenerator.generate_pod_targets(resolver_specs_by_target, target_inspections, sandbox, installation_options)
       end
 
       # Generates dependencies that require the specific version of the Pods
@@ -830,6 +525,13 @@ module Pod
         end
       end
 
+      # Fetches an external source and stores it into the sandbox
+      #
+      # @param [Pod::Dependency] dependency the dependency to fetch
+      #
+      # @param [Boolean] use_lockfile_options whether to use the lockfile's checkout options
+      #        when fetching the dependency
+      #
       def fetch_external_source(dependency, use_lockfile_options)
         source = if use_lockfile_options && lockfile && checkout_options = lockfile.checkout_options_for_pod_named(dependency.root_name)
                    ExternalSources.from_params(checkout_options, dependency, podfile.defined_in_file, installation_options.clean?)
@@ -839,6 +541,11 @@ module Pod
         source.fetch(sandbox)
       end
 
+      # @param [SpecsState] podfile_state
+      #         the state of the Podfile specs
+      #
+      # @return [Array<Pod::Dependency>] the external dependencies to be fetched
+      #
       def dependencies_to_fetch(podfile_state)
         @deps_to_fetch ||= begin
           deps_to_fetch = []
@@ -860,6 +567,8 @@ module Pod
         end
       end
 
+      # @return [Boolean]
+      #
       def checkout_requires_update?(dependency)
         return true unless lockfile && sandbox.manifest
         locked_checkout_options = lockfile.checkout_options_for_pod_named(dependency.root_name)
@@ -867,6 +576,8 @@ module Pod
         locked_checkout_options != sandbox_checkout_options
       end
 
+      # @return [Array<Pod::Depedency>]
+      #
       def pods_to_fetch(podfile_state)
         @pods_to_fetch ||= begin
           pods_to_fetch = podfile_state.added + podfile_state.changed
@@ -882,6 +593,10 @@ module Pod
         end
       end
 
+      # Stores the existing checkout options for external sources into the sandbox
+      #
+      # @return [void]
+      #
       def store_existing_checkout_options
         podfile_dependencies.select(&:external_source).each do |dep|
           if checkout_options = lockfile && lockfile.checkout_options_for_pod_named(dep.root_name)
@@ -890,7 +605,7 @@ module Pod
         end
       end
 
-      # Converts the Podfile in a list of specifications grouped by target.
+      # Converts the Podfile into a list of specifications grouped by target.
       #
       # @note   As some dependencies might have external sources the resolver
       #         is aware of the {Sandbox} and interacts with it to download the
@@ -906,7 +621,7 @@ module Pod
       # @note   In update mode the resolver is set to always update the specs
       #         from external sources.
       #
-      # @return [Hash{TargetDefinition => Array<Spec>}] the specifications
+      # @return [Hash{TargetDefinition => Array<Specification>}] the specifications
       #         grouped by target.
       #
       def resolve_dependencies(locked_dependencies)
@@ -974,25 +689,6 @@ module Pod
         sandbox_state
       end
 
-      # @param  [Platform] platform
-      #         The platform to build against
-      #
-      # @return [Boolean] Whether the platform requires 64-bit architectures
-      #
-      def requires_64_bit_archs?(platform)
-        return false unless platform
-        case platform.name
-        when :osx
-          true
-        when :ios
-          platform.deployment_target >= IOS_64_BIT_ONLY_VERSION
-        when :watchos
-          false
-        when :tvos
-          false
-        end
-      end
-
       #-----------------------------------------------------------------------#
 
       # @!group Analysis sub-steps
@@ -1013,16 +709,18 @@ module Pod
 
       # Precompute information for each target_definition in the Podfile
       #
+      # @param [Array<TargetDefinition>] target_definitions the targets to inspect
+      #
       # @note The platforms are computed and added to each target_definition
       #       because it might be necessary to infer the platform from the
       #       user targets.
       #
       # @return [Hash{TargetDefinition => TargetInspectionResult}]
       #
-      def inspect_targets_to_integrate
+      def inspect_targets_to_integrate(target_definitions)
         inspection_result = {}
         UI.section 'Inspecting targets to integrate' do
-          inspectors = @podfile_dependency_cache.target_definition_list.map do |target_definition|
+          inspectors = target_definitions.map do |target_definition|
             next if target_definition.abstract?
             TargetInspector.new(target_definition, config.installation_root)
           end.compact
